feat: wip creating self element
